add map.count to check if area of map exists, fixed pickups, fixed styling

+++ src/game/GameManager.cpp
@@ -98,14 +98,16 @@ bool GameManager::createMarine(SDL_Renderer* gRenderer, const float x, const flo
 
     Marine m(id, marineRect, moveRect, projRect, damRect);
     marineManager.insert({id, m});
-
-    if (!marineManager.at(id).texture.loadFromFile("assets/texture/arrow.png", gRenderer)) {
-        logv("Failed to load the marine texture!\n");
-        deleteMarine(id);
-        return false;
+    if(marineManager.count(id)){
+        if (!marineManager.at(id).texture.loadFromFile("assets/texture/arrow.png", gRenderer)) {
+            logv("Failed to load the marine texture!\n");
+            deleteMarine(id);
+            return false;
+        }
+        marineManager.at(id).setPosition(x,y);
+        return true;
     }
-    marineManager.at(id).setPosition(x,y);
-    return true;
+    return false;
 }
 
 void GameManager::deleteMarine(const int32_t id) {
@@ -164,14 +166,16 @@ int32_t GameManager::createTurret(SDL_Renderer* gRenderer, const float x, const 
     SDL_Rect pickRect = {0,0,100,100};
 
     turretManager.insert({id, Turret(id, turretRect, moveRect, projRect, damRect, pickRect)});
-
-    if (!turretManager.at(id).texture.loadFromFile("assets/texture/turret.png", gRenderer)) {
-        logv("Failed to load the turret texture!\n");
-        deleteTurret(id);
-        return -1;
+    if(turretManager.count(id)){
+        if (!turretManager.at(id).texture.loadFromFile("assets/texture/turret.png", gRenderer)) {
+            logv("Failed to load the turret texture!\n");
+            deleteTurret(id);
+            return -1;
+        }
+        turretManager.at(id).setPosition(x,y);
+        return id;
     }
-    turretManager.at(id).setPosition(x,y);
-    return id;
+    return -1;
 }
 
 // Get a tower by its id
@@ -197,15 +201,16 @@ bool GameManager::createZombie(SDL_Renderer* gRenderer, const float x, const flo
 
 
     zombieManager.insert({id, Zombie(id, zombieRect, moveRect, projRect, damRect)});
-
-    if (!zombieManager.at(id).texture.loadFromFile("assets/texture/zombie.png", gRenderer)) {
-        logv("Failed to load the Zombie texture!\n");
-        deleteZombie(id);
-        return false;
+    if(zombieManager.count(id)){
+        if (!zombieManager.at(id).texture.loadFromFile("assets/texture/zombie.png", gRenderer)) {
+            logv("Failed to load the Zombie texture!\n");
+            deleteZombie(id);
+            return false;
+        }
+        zombieManager.at(id).setPosition(x,y);
+        return true;
     }
-    zombieManager.at(id).setPosition(x,y);
-
-    return true;
+    return false;
 }
 
 // Deletes zombie from level
@@ -229,9 +234,11 @@ int32_t GameManager::addWeapon(std::shared_ptr<Weapon> weapon){
 
     const int32_t id = weapon->getId();
     weaponManager.insert({id, weapon});
-    weaponManager.at(id)->setId(id);
-
-    return id;
+    if(weaponManager.count(id)){
+        weaponManager.at(id)->setId(id);
+        return id;
+    }
+    return -1;
 
 }
 
@@ -251,35 +258,38 @@ bool GameManager::createWeaponDrop(SDL_Renderer* gRenderer, const float x, const
     const int32_t wid = w.getId();
     const int32_t id = generateID();
 
-    SDL_Rect weaponDropRect = {0,0,100,100};
-    SDL_Rect pickRect = {0,0,100,100};
+    SDL_Rect weaponDropRect = {x,y,100,100};
+    SDL_Rect pickRect = {x,y,100,100};
 
     addWeapon(std::dynamic_pointer_cast<Weapon>(std::make_shared<Rifle>(w)));
 
     WeaponDrop wd(id, weaponDropRect, pickRect, wid);
     weaponDropManager.insert({id, wd});
 
-    weaponDropManager.at(id).setX(x);
-    weaponDropManager.at(id).setY(y);
+    if(weaponDropManager.count(id)){
+        weaponDropManager.at(id).setX(x);
+        weaponDropManager.at(id).setY(y);
 
-    if(!weaponDropManager.at(id).texture.loadFromFile("assets/texture/shotGun.png", gRenderer)) {
-        logv("Failed to load the player texture!\n");
-        deleteWeaponDrop(id);
-        return false;
+        if(!weaponDropManager.at(id).texture.loadFromFile("assets/texture/shotGun.png", gRenderer)) {
+            logv("Failed to load the player texture!\n");
+            deleteWeaponDrop(id);
+            return false;
+        }
+        weaponDropManager.at(id).setPosition(x,y);
+        return true;
     }
-    weaponDropManager.at(id).setPosition(x,y);
-    return true;
+    return false;
 }
 
 //returns weapon drop in  weaponDropManager
 WeaponDrop& GameManager::getWeaponDrop(const int32_t id){
-    return weaponDropManager.at(id);
+        return weaponDropManager.at(id);
 }
 
 //returns weapon in weaponManager
 std::shared_ptr<Weapon> GameManager::getWeapon(const int32_t id){
-    const auto& it = weaponManager.find(id);
-    if(it != weaponManager.end()){
+
+    if(weaponManager.count(id)){
         return weaponManager.at(id);
     }
     logv("Couldnt find Weapon\n");
@@ -350,13 +360,16 @@ int32_t GameManager::createBarricade(SDL_Renderer* gRenderer, const float x, con
 
     Barricade b(id, barricadeRect, moveRect, pickRect);
     barricadeManager.insert({id, b});
-    if (!barricadeManager.at(id).texture.loadFromFile("assets/texture/barricade.png", gRenderer)) {
-        logv("Failed to load the barricade texture!\n");
-        deleteBarricade(id);
-        return -1;
+    if(barricadeManager.count(id)){
+        if (!barricadeManager.at(id).texture.loadFromFile("assets/texture/barricade.png", gRenderer)) {
+            logv("Failed to load the barricade texture!\n");
+            deleteBarricade(id);
+            return -1;
+        }
+        barricadeManager.at(id).setPosition(x,y);
+        return id;
     }
-    barricadeManager.at(id).setPosition(x,y);
-    return id;
+    return -1;
 }
 
 
@@ -380,16 +393,18 @@ int32_t GameManager::createWall(SDL_Renderer* gRenderer,
 
     wallManager.insert({id, Wall(id, wallRect, moveRect, pickRect, h, h)});
 
-    if (!wallManager.at(id).texture.loadFromFile("assets/texture/wall.png", gRenderer)) {
-        logv("Failed to load the wall texture!\n");
-        deleteBarricade(id);
-        return -1;
+    if(wallManager.count(id)){
+        if (!wallManager.at(id).texture.loadFromFile("assets/texture/wall.png", gRenderer)) {
+            logv("Failed to load the wall texture!\n");
+            deleteBarricade(id);
+            return -1;
+        }
+
+        wallManager.at(id).texture.setDimensions(w, h);
+        wallManager.at(id).setPosition(x,y);
+        return id;
     }
-
-    wallManager.at(id).texture.setDimensions(w, h);
-    wallManager.at(id).setPosition(x,y);
-
-    return id;
+    return -1;
 }
